Deprecates url_replace template tags.
Adds deprecation warnings to the url_replace and
url_replace_diff template tags, recommending the use of
the querystring template tag instead.

Updates dependencies in poetry.lock, including Django to 5.2.4,
cairosvg to 2.8.2, certifi to 2025.6.15, charset-normalizer to 3.4.2,
click to 8.2.1, exceptiongroup to 1.3.0, markdown to 3.8.2,
packaging to 25.0, pillow to 11.3.0, platformdirs to 4.3.8,
pluggy to 1.6.0, pygments to 2.19.2, pyyaml-env-tag to 1.1,
requests to 2.32.4, urllib3 to 2.5.0 and tomli (conditional dependency)
and typing-extensions (conditional dependency).